backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
PyMuPDF==1.24.5
openai==1.35.3
python-dotenv==1.0.1
groq
httpx